+++ contao/dca/tl_user.php
@@ -1,7 +1,13 @@
 <?php
 
-$GLOBALS['TL_DCA']['tl_user']['palettes']['extend'] = str_replace('newsfeedp;', 'newsfeedp;{locations_legend},locations,locations_default;', $GLOBALS['TL_DCA']['tl_user']['palettes']['extend']);
-$GLOBALS['TL_DCA']['tl_user']['palettes']['custom'] = str_replace('newsfeedp;', 'newsfeedp;{locations_legend},locations,locations_default;', $GLOBALS['TL_DCA']['tl_user']['palettes']['custom']);
+use Contao\CoreBundle\DataContainer\PaletteManipulator;
+
+PaletteManipulator::create()
+    ->addLegend('locations_legend', 'filemounts_legend')
+    ->addField('locations', 'locations_legend', PaletteManipulator::POSITION_APPEND)
+    ->addField('locations_default', 'locations_legend', PaletteManipulator::POSITION_APPEND)
+    ->applyToPalette('extend', 'tl_user')
+    ->applyToPalette('custom', 'tl_user');
 
 $GLOBALS['TL_DCA']['tl_user']['fields']['locations'] = [
     'label'     => &$GLOBALS['TL_LANG']['tl_user']['locations'],