+++ contao/dca/tl_user_group.php
@@ -1,8 +1,14 @@
 <?php
 
+use Contao\CoreBundle\DataContainer\PaletteManipulator;
+
 Contao\System::loadLanguageFile('tl_user');
 
-$GLOBALS['TL_DCA']['tl_user_group']['palettes']['default'] = str_replace('newsfeedp;', 'newsfeedp;{locations_legend},locations,locations_default;', $GLOBALS['TL_DCA']['tl_user_group']['palettes']['default']);
+PaletteManipulator::create()
+    ->addLegend('locations_legend', 'filemounts_legend')
+    ->addField('locations', 'locations_legend', PaletteManipulator::POSITION_APPEND)
+    ->addField('locations_default', 'locations_legend', PaletteManipulator::POSITION_APPEND)
+    ->applyToPalette('default', 'tl_user_group');
 
 $GLOBALS['TL_DCA']['tl_user_group']['fields']['locations'] = [
     'label'     => &$GLOBALS['TL_LANG']['tl_user']['locations'],