+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased] - 2022-04-11
+- Fixed: user and user group palettes
+
 ## [2.0.1] - 2021-11-01
 - Fixed: missing translations for config element type
 - Fixed: return value of config element type not formatted